split play mailer into two modules and makes it not depend on play
actually this will change Module to AbstractModule and remove
Configuration and Logger from play-mailer so that it does not
depend directly on play, this makes it easier to use play-mailer
without play and on newer play versions
