+++ play-mailer/src/main/java/play/libs/mailer/Attachment.java
