+++ play-mailer/src/test/resources/play_icon_full_color.png
