+++ samples/runtimeDI/build.sbt
@@ -7,10 +7,9 @@ version := "1.0-SNAPSHOT"
 
 scalaVersion := "2.12.1"
 
-resolvers += "Scalaz Bintray Repo" at "http://dl.bintray.com/scalaz/releases"
-
 libraryDependencies ++= Seq(
-  "com.typesafe.play" %% "play-mailer" % "5.0.1-SNAPSHOT"
+  "com.typesafe.play" %% "play-mailer" % "5.0.1-SNAPSHOT",
+  "com.typesafe.play" %% "play-mailer-guice" % "5.0.1-SNAPSHOT",
 )
 
 lazy val root = (project in file(".")).enablePlugins(PlayJava)
