--- src/main/resources/reference.conf
@@ -1,23 +1,0 @@
-play {
-  modules {
-    enabled += "play.api.libs.mailer.MailerModule"
-    enabled += "play.api.libs.mailer.SMTPConfigurationModule"
-  }
-  mailer {
-    host = null
-    port = 25
-    ssl = no
-    tls = no
-    tlsRequired = no
-    user = null
-    password = null
-    // Defaults to no, to take effect you also need to set the log level to "DEBUG" for the application logger
-    debug = no
-    // Defaults to no, will only log all the email properties instead of sending an email
-    mock = no
-    // Set the socket I/O timeout value in milliseconds. Default is 60 second timeout.
-    timeout = null
-    // Set the socket connection timeout value in milliseconds. Default is a 60 second timeout.
-    connectiontimeout = null
-  }
-}
